pyproject.toml
[project]
name = "fermax-blue-hass"
version = "0.1.0"
description = "Home Assistant integration for Fermax Blue intercoms"
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-homeassistant-custom-component>=0.13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
